+++ iceoryx_posh/include/iceoryx_posh/roudi/roudi_config.hpp
@@ -27,7 +27,7 @@ struct RouDiConfig
     struct
     {
         /// @todo since we are now fully relocatable, this can be removed
-        uintptr_t m_sharedMemoryBaseAddressOffset = 0;
+        uint64_t m_sharedMemoryBaseAddressOffset = 0;
         bool m_verifySharedMemoryPlacement = false;
     } roudi;
